fix moved page plugin to separate module
