+++ crates/seo-plugins/src/plugins/axe.rs
@@ -1,7 +1,8 @@
 use std::any::Any;
 
 use crate::utils::{
-    config::{CheckResult, Rule, RuleCategory, SeoPlugin, Severity},
+    config::{CheckResult, Rule, RuleCategory, Severity},
+    page_plugin::SeoPlugin,
     registry::PluginRegistry,
 };
 use scraper::Selector;
@@ -520,10 +521,7 @@ fn get_required_aria_attributes(role: &str) -> Vec<&'static str> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::utils::{
-        config::{RuleConfig, SeoPlugin},
-        page::Page,
-    };
+    use crate::utils::{config::RuleConfig, page::Page, page_plugin::SeoPlugin};
     use hyper::service::{make_service_fn, service_fn};
     use hyper::{Body, Response, Server};
