+++ crates/seo-plugins/src/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod config;
 pub mod page;
+pub mod page_plugin;
 pub mod registry;
