+++ turbo/generators/templates/plugin.rs.hbs
@@ -1,7 +1,8 @@
 use std::any::Any;
 
 use crate::utils::{
-    config::{CheckResult, Rule, RuleCategory, SeoPlugin, Severity},
+    config::{CheckResult, Rule, RuleCategory, Severity},
+    page_plugin::SeoPlugin,
     registry::PluginRegistry,
 };
 
@@ -69,8 +70,9 @@ impl SeoPlugin for {{ pluginName }}Plugin {
 mod tests {
     use super::*;
     use crate::utils::{
-        config::{RuleConfig, SeoPlugin},
+        config::RuleConfig,
         page::Page,
+        page_plugin::SeoPlugin,
     };
     use hyper::service::{make_service_fn, service_fn};
     use hyper::{Body, Response, Server};
